requirements.txt
fastapi>=0.110.0
uvicorn>=0.27.0
python-dotenv>=1.0.0
pydantic>=2.6.0
httpx>=0.27.0
playwright>=1.41.0
ffmpeg-python>=0.2.0
prisma>=0.12.0
python-multipart>=0.0.9
jinja2>=3.1.0
requests>=2.31.0
Pillow>=10.2.0
aiofiles>=23.2.0
google-generativeai>=0.4.0
openai>=1.12.0
boto3>=1.34.0
imageio-ffmpeg>=0.4.9
